More work on https://github.com/wwivbbs/wwiv/issues/238
