+++ bbs/asv.cpp
@@ -20,9 +20,14 @@
 
 #include "bbs/datetime.h"
 #include "bbs/bbs.h"
+#include "bbs/bputch.h"
+#include "bbs/com.h"
 #include "bbs/fcns.h"
 #include "bbs/inmsg.h"
 #include "bbs/input.h"
+#include "bbs/pause.h"
+#include "bbs/sysopf.h"
+#include "bbs/sysoplog.h"
 #include "bbs/vars.h"
 #include "bbs/wsession.h"
 #include "bbs/keycodes.h"
