+++ bbs/bbs_lib.vcxproj
@@ -91,10 +91,16 @@
     <ClInclude Include="automsg.h" />
     <ClInclude Include="bbs.h" />
     <ClInclude Include="bbslist.h" />
+    <ClInclude Include="bbsutl.h" />
+    <ClInclude Include="bbsutl1.h" />
+    <ClInclude Include="bgetch.h" />
+    <ClInclude Include="bputch.h" />
     <ClInclude Include="callback.h" />
     <ClInclude Include="capture.h" />
+    <ClInclude Include="com.h" />
     <ClInclude Include="common.h" />
     <ClInclude Include="conf.h" />
+    <ClInclude Include="connect1.h" />
     <ClInclude Include="crc.h" />
     <ClInclude Include="datetime.h" />
     <ClInclude Include="dropfile.h" />
@@ -104,6 +110,7 @@
     <ClInclude Include="inmsg.h" />
     <ClInclude Include="input.h" />
     <ClInclude Include="instmsg.h" />
+    <ClInclude Include="interpret.h" />
     <ClInclude Include="keycodes.h" />
     <ClInclude Include="listplus.h" />
     <ClInclude Include="local_io.h" />
@@ -112,6 +119,7 @@
     <ClInclude Include="menuspec.h" />
     <ClInclude Include="menusupp.h" />
     <ClInclude Include="menu_parser.h" />
+    <ClInclude Include="msgbase.h" />
     <ClInclude Include="net.h" />
     <ClInclude Include="new_bbslist.h" />
     <ClInclude Include="null_local_io.h" />
@@ -128,6 +136,7 @@
     <ClInclude Include="qwk.h" />
     <ClInclude Include="runnable.h" />
     <ClInclude Include="subxtr.h" />
+    <ClInclude Include="sysopf.h" />
     <ClInclude Include="sysoplog.h" />
     <ClInclude Include="template.h" />
     <ClInclude Include="utility.h" />
