+++ bbs/fcns.h
@@ -80,37 +80,11 @@ void ListAllColors();
 
 // File: bbsutl.cpp
 
-bool inli(std::string* outBuffer, std::string* rollOver, std::string::size_type nMaxLen, bool bAddCRLF = true,
-          bool bAllowPrevious = false, bool bTwoColorChatMode = false, bool clear_previous_line = false);
-bool inli(char *pszBuffer, char *pszRollover, std::string::size_type nMaxLen, bool bAddCRLF = true,
-          bool bAllowPrevious = false, bool bTwoColorChatMode = false, bool clear_previous_line = false);
-bool so();
-bool cs();
-bool lcs();
-bool checka();
-bool checka(bool *abort);
-bool checka(bool *abort, bool *next);
-void pla(const std::string& text, bool *abort);
-void plal(const std::string& text, std::string::size_type limit, bool *abort);
-bool sysop2();
-bool checkcomp(const char *pszComputerType);
-int  check_ansi();
-bool set_language_1(int n);
-bool set_language(int n);
-// todo(rush): make this a C++11 enum
-char *mmkey(int dl, int area = 0 /* mmkeyNoArea */, bool bListOption = false);
-const char *YesNoString(bool bYesNo);
+#include "bbs/bbsutl.h"
 
 
 // File: bbsutl1.cpp
-
-bool AllowLocalSysop();
-void parse_email_info(const std::string& emailAddress, int *pUserNumber, int *pSystemNumber);
-bool ValidateSysopPassword();
-void hang_it_up();
-bool play_sdf(const std::string& soundFileName, bool abortable);
-std::string describe_area_code(int nAreaCode);
-std::string describe_area_code_prefix(int nAreaCode, int town);
+#include "bbs/bbsutl1.h"
 
 
 // File: bbsutl2.cpp
@@ -123,19 +97,10 @@ std::string strip_to_node(const std::string& txt);
 
 
 // File: bgetch.cpp
-
-char bgetch();
-char bgetchraw();
-bool bkbhitraw();
-bool bkbhit();
-
+#include "bbs/bgetch.h"
 
 // File: bputch.cpp
-
-int  bputch(char c, bool bUseInternalBuffer = false);
-void FlushOutComChBuffer();
-void rputch(char ch, bool bUseInternalBuffer = false);
-void rputs(const char *pszText);
+#include "bbs/bputch.h"
 
 // File: chains.cpp
 
@@ -164,16 +129,7 @@ void buildcolorfile();
 
 
 // File: com.cpp
-void RestoreCurrentLine(const char *cl, const char *atr, const char *xl, const char *cc);
-void dump();
-bool CheckForHangup();
-void makeansi(int attr, char *pszOutBuffer, bool forceit);
-void resetnsp();
-char getkey();
-bool yesno();
-bool noyes();
-char ynq();
-char onek(const char *pszAllowableChars, bool bAutoMpl = false);
+#include "bbs/com.h"
 
 
 // File: conf.cpp
@@ -210,16 +166,7 @@ void changedsl();
 
 
 // File: connect1.cpp
-
-void zap_call_out_list();
-void read_call_out_list();
-void zap_bbs_list();
-void read_bbs_list_index();
-bool valid_system(int ts);
-net_system_list_rec *next_system(int ts);
-void zap_contacts();
-void read_contacts();
-void set_net_num(int nNetworkNumber);
+#include "bbs/connect1.h"
 
 // File: defaults.cpp
 
@@ -279,7 +226,6 @@ int finduser1(const std::string& searchString);
 
 // File: gfiles.cpp
 
-char *get_file(const std::string& filename, long *len);
 gfilerec *read_sec(int sn, int *nf);
 void gfiles();
 
@@ -321,11 +267,6 @@ void instance_edit();
 void write_inst(int loc, int subloc, int flags);
 bool inst_msg_waiting();
 int  setiia(int poll_ticks);
-
-
-// File: interpret.cpp
-
-const char *interpret(char chKey);
 
 
 // File: lilo.cpp
@@ -396,24 +337,7 @@ void Packers();
 
 
 // File: msgbase.cpp
-
-void remove_link(messagerec * pMessageRecord, const std::string fileName);
-void savefile(char *b, long lMessageLength, messagerec * pMessageRecord, const std::string fileName);
-char *readfile(messagerec * pMessageRecord, const std::string fileName, long *plMessageLength);
-void LoadFileIntoWorkspace(const char *pszFileName, bool bNoEditAllowed);
-bool ForwardMessage(int *pUserNumber, int *pSystemNumber);
-std::unique_ptr<File> OpenEmailFile(bool bAllowWrite);
-void sendout_email(const std::string& title, messagerec * msg, int anony, int nUserNumber, int nSystemNumber, int an,
-                   int nFromUser, int nFromSystem, int nForwardedCode, int nFromNetworkNumber);
-bool ok_to_mail(int nUserNumber, int nSystemNumber, bool bForceit);
-void email(int nUserNumber, int nSystemNumber, bool forceit, int anony, bool force_title = false,
-           bool bAllowFSED = true);
-void imail(int nUserNumber, int nSystemNumber);
-void read_message1(messagerec * pMessageRecord, char an, bool readit, bool *next, const char *pszFileName,
-                   int nFromSystem, int nFromUser);
-void read_message(int n, bool *next, int *val);
-void lineadd(messagerec* pMessageRecord, const std::string& sx, const std::string fileName);
-
+#include "bbs/msgbase.h"
 
 // File: msgbase1.cpp
 
@@ -593,19 +517,7 @@ void chat1(char *pszChatLine, bool two_way);
 // File: sysoplog.cpp
 #include "sysoplog.h"
 // File: sysopf.cpp
-
-void reset_files();
-void prstatus();
-void valuser(int nUserNumber);
-void print_net_listing(bool bForcePause);
-void read_new_stuff();
-void mailr();
-void chuser();
-void zlog();
-void auto_purge();
-void beginday(bool displayStatus);
-void set_user_age();
-
+#include "bbs/sysopf.h"
 
 // File: uedit.cpp
 
@@ -690,7 +602,7 @@ void arc_l();
 
 // File: xferovl1.cpp
 
-void modify_extended_description(char **sss, const char *dest, const char *title);
+void modify_extended_description(char **sss, const char *dest);
 bool valid_desc(const char *pszDescription);
 bool get_file_idz(uploadsrec * pUploadRecord, int dn);
 int  read_idz_all();
