+++ bbs/new_bbslist.cpp
@@ -28,9 +28,12 @@
 #include <vector>
 
 #include "bbs/bbs.h"
-#include "bbs/fcns.h"
+#include "bbs/bbsutl.h"
+#include "bbs/com.h"
 #include "bbs/input.h"
+#include "bbs/pause.h"
 #include "bbs/printfile.h"
+#include "bbs/sysopf.h"
 #include "bbs/vars.h"  // for syscfg
 #include "bbs/wsession.h"
 #include "core/stl.h"
