+++ bbs/platform/win32/wiot.cpp
@@ -24,7 +24,6 @@
 
 #include "bbs/wuser.h"
 #include "bbs/platform/platformfcns.h"
-#include "bbs/fcns.h"
 #include "core/strings.h"
 #include "core/file.h"
 #include "core/wwivport.h"
