+++ bbs/smallrecord.cpp
@@ -19,6 +19,7 @@
 #include <iostream>
 
 #include "bbs/bbs.h"
+#include "bbs/sysoplog.h"
 #include "bbs/vars.h"
 #include "bbs/wconstants.h"
 #include "bbs/wuser.h"
@@ -27,9 +28,6 @@
 #include "core/strings.h"
 #include "core/wwivport.h"
 #include "sdk/filenames.h"
-
-// TODO - Remove this and finduser, finduser1, ISR, DSR, and add_add
-#include "bbs/fcns.h"
 
 using namespace wwiv::strings;
