+++ bbs/subacc.cpp
@@ -20,7 +20,9 @@
 #include <string>
 
 #include "bbs/bbs.h"
-#include "bbs/fcns.h"
+#include "bbs/bbsutl.h"
+#include "bbs/connect1.h"
+#include "bbs/msgbase.h"
 #include "bbs/subxtr.h"
 #include "bbs/pause.h"
 #include "sdk/vardec.h"
